Separate components for legacy and new demo charts
